Use a random unused port.
In situations where supertest is run multiple times in parallel on the
same host EADDRINUSE can result. This is due to the current method of
using port 3456 (+ offset). A more reliable method for obtaining a
random unused port is to specify port 0 and have the OS choose which
port to bind.

Fixes #62