add expect(function(res) {}) syntax
The ability to add a function assertion makes it easy to create
reusable assertions (and simple plugins). For instance, an API
might require a certain content-type and 'next' and 'previous'
keys in a number of endpoints. With the ability to pass in an
assertion function this usecase is very simple:

  .expect(validPagedApiResponse);

  function validPagedApiResponse(res) {
    if(!('next' in res.body)) return "All responses require a 'next'
    key";
  }