Modifies expect to not throw error unless real error #144
The previous behaviour of expect was to throw any returned truthy values
as an error, to be passed to the subsequent end callback.

This presents problems in coffeescript due to implicit returns,
requiring an explicit `return` with a falsy value to prevent triggering
a failure.

This commit changes this behaviour so that `expect` will only throw the
truthy value forward if it is an instance of `Error`. This seems to make
move sense anyway, where truthy values don't typically indicate a
failure.